requirements.txt
aiosqlite==0.17.0
anyio==4.0.0
async-timeout==4.0.3
certifi==2023.7.22
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iso8601==1.1.0
multidict==6.0.4
pypika-tortoise==0.1.6
pytz==2023.3
redis==5.0.0
sniffio==1.3.0
tortoise-orm==0.20.0
typing_extensions==4.7.1
yarl==1.9.2
